requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
urllib3==2.6.0
whitenoise==6.11.0
